LIBITD-1550. Refactored "calculate_relative_path" into "utils.py"
Moved "calculate_relative_path" out of the "Assets" class into
"utils.py".

https://issues.umd.edu/browse/LIBITD-1550
